discovery/scanner/requirements.txt
azure-identity>=1.17.0
azure-mgmt-resourcegraph>=8.0.0
azure-cosmos>=4.7.0
